app/pyproject.toml
[project]
name = "model-router"
version = "1.0.0"
description = "Provider-agnostic LLM router with cost optimisation, async providers, caching, circuit breaker, and Prometheus/Grafana observability"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.1",
    "uvicorn[standard]==0.30.1",
    "pydantic==2.7.4",
    "httpx==0.27.0",
    "PyYAML==6.0.1",
    "prometheus-client==0.20.0",
    "google-genai==0.8.0",
    "slowapi==0.1.9",
    "structlog==24.2.0",
    "opentelemetry-sdk==1.24.0",
    "opentelemetry-instrumentation-fastapi==0.45b0",
    "opentelemetry-instrumentation-httpx==0.45b0",
    "redis[asyncio]==5.0.7",
    "watchfiles==0.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "pytest-mock==3.14.0",
    "ruff==0.4.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
